The patch below adds support for the 2.6.29 kernel and also fixes the checkpatch issues reported by the checkpatch script included with the 2.6.29 kernel and that were not yet reported by the 2.6.28 checkpatch script (trailing statements should be on the next line / struct should normally be const). The patch below has been tested as follows: - Reran scripts/run-regression-tests -k 2.6.24.7 -k 2.6.25.20 -k 2.6.26.8 -k 2.6.27.21 -k 2.6.28.9 -k 2.6.29 and verified the output. - Rebuilt, installed and loaded scst, iscsi-scst and srpt as follows: make -s clean && make -s -C scst install && make -s -C iscsi-scst install && make -s -C srpt install && cd scstadmin && make -s && make -s install && modprobe scst_vdisk && modprobe iscsi-scst && dmesg

with minor cleanups and corrections in put_page_callback-2.6.29.patch



git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@717 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ iscsi-scst/kernel/iscsi.c
@@ -3156,7 +3156,7 @@ static int __init iscsi_init(void)
 	if (err < 0)
 		goto out_reg_tmpl;
 
-	num = max(num_online_cpus(), 2);
+	num = max((int)num_online_cpus(), 2);
 
 	err = iscsi_run_threads(num, "iscsird", istrd);
 	if (err != 0)
